Use Webpack.
