+++ package.json
@@ -4,6 +4,8 @@
     "description": "",
     "main": "index.js",
     "scripts": {
+        "dev": "webpack-dev-server --open --mode development",
+        "build": "webpack --mode production",
         "test": "echo \"Error: no test specified\" && exit 1",
         "lint:js": "eslint .",
         "lint:css": "stylelint **/*.{css,scss}",
@@ -12,12 +14,24 @@
     "author": "Osama Aldemeery <aldemeery@gmail.com>",
     "license": "MIT",
     "devDependencies": {
-        "eslint": "^7.4.0",
-        "eslint-config-airbnb": "^18.2.0",
+        "@babel/core": "^7.10.5",
+        "@babel/preset-env": "^7.10.4",
+        "@babel/plugin-proposal-class-properties": "^7.10.4",
+        "babel-eslint": "^10.1.0",
+        "babel-loader": "^8.1.0",
+        "css-loader": "^4.1.1",
+        "eslint": "^7.5.0",
+        "eslint-config-airbnb-base": "^14.2.0",
         "eslint-plugin-import": "^2.22.0",
+        "html-webpack-plugin": "^4.3.0",
+        "style-loader": "^1.2.1",
         "stylelint": "^13.6.1",
         "stylelint-config-standard": "^20.0.0",
         "stylelint-csstree-validator": "^1.8.0",
-        "stylelint-scss": "^3.18.0"
+        "stylelint-scss": "^3.18.0",
+        "url-loader": "^4.1.0",
+        "webpack": "^4.44.1",
+        "webpack-cli": "^3.3.12",
+        "webpack-dev-server": "^3.11.0"
     }
 }
