+++ src/index.html
@@ -8,7 +8,6 @@
             rel="stylesheet"
             href="https://cdn.jsdelivr.net/gh/kognise/water.css@latest/dist/dark.min.css"
         />
-        <link rel="stylesheet" href="./style.css" />
     </head>
     <body>
         <h1>ToDo <small class="current-project">(Default)</small></h1>
@@ -73,6 +72,5 @@
             </thead>
             <tbody id="body"></tbody>
         </table>
-        <script src="./scripts.js"></script>
     </body>
 </html>
